[#8] Move KafkaProducerConfigProperties to module kafka-common.
The configuration class was in hono-core (like the ClientConfigProperties),
but it is better placed in the kafka-common module. Now it can also depend on
Apache Kafka client.
