+++ service-base/src/main/java/org/eclipse/hono/service/AbstractAdapterConfig.java
@@ -43,11 +43,11 @@ import org.eclipse.hono.client.SendMessageSampler;
 import org.eclipse.hono.config.ApplicationConfigProperties;
 import org.eclipse.hono.config.AuthenticatingClientConfigProperties;
 import org.eclipse.hono.config.ClientConfigProperties;
-import org.eclipse.hono.config.KafkaProducerConfigProperties;
 import org.eclipse.hono.config.ProtocolAdapterProperties;
 import org.eclipse.hono.config.ServerConfig;
 import org.eclipse.hono.config.VertxProperties;
 import org.eclipse.hono.kafka.client.CachingKafkaProducerFactory;
+import org.eclipse.hono.kafka.client.KafkaProducerConfigProperties;
 import org.eclipse.hono.service.cache.SpringCacheProvider;
 import org.eclipse.hono.service.conditions.ConditionalOnAmqpMessaging;
 import org.eclipse.hono.service.conditions.ConditionalOnKafkaMessaging;
